Use relative paths in redirects
When the UI is served behind a reverse proxy (e.g. traefik), the pages
may be served under a subpath. This breaks our redirects and calls to
the backend API. In order to overcome this we use relative paths
everywhere.

This is not the proper solution and ideally we would want the reverse
proxy to inject the html with the base URL tag.